fix network test

+++ extensions/net/main.cpp
@@ -27,7 +27,7 @@ static int mlua_send(lua_State* L)
 static int mlua_receive(lua_State* L)
 {
 	reception_buffer[0] = 0; // assure that even if Network doesn't touch the buffer, it's still valid
-	int received_or_error = net.receive(reception_buffer, sizeof(reception_buffer));
+	long received_or_error = net.receive(reception_buffer, sizeof(reception_buffer));
 	lua_pushnumber(L, received_or_error);
 	if (received_or_error >= 0) {
 		reception_buffer[received_or_error] = 0;
